refactor for test and readme
