+++ src/main/java/com/trading/app/component/SignalHandler.java
@@ -1,4 +1,4 @@
-package com.trading.app.util;
+package com.trading.app.component;
 
 /**
  * This is an upcall from our trading system, and we cannot change it.
