+++ src/test/java/com/trading/app/component/SingnalHandlerTest.java
@@ -1,13 +1,12 @@
-package com.trading.app.util;
+package com.trading.app.component;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.trading.app.dto.Action;
 import com.trading.app.dto.SignalRequest;
 import com.trading.app.entity.Signal;
 import com.trading.app.exception.ResourceNotFoundException;
-import com.trading.app.lib.Algo;
-import com.trading.app.service.SignalService;
+import com.trading.app.service.SignalServiceImpl;
+import com.trading.app.util.ObjectMapperUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
@@ -19,32 +18,31 @@ import org.springframework.test.context.ActiveProfiles;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.when;
 
 @ActiveProfiles("test")
 @SpringBootTest
 public class SingnalHandlerTest {
     @Mock
-    private SignalService signalService;
+    private SignalServiceImpl signalService;
     @Autowired
     private ObjectMapper objectMapper;
-    private Algo algo;
+
     private SignalHandlerImpl signalHandler;
 
-     @BeforeEach
+    @BeforeEach
     public void setup() {
         signalHandler = new SignalHandlerImpl(signalService, objectMapper);
-        algo = new Algo();
     }
 
     @Test
     public void testHandleSignalWithValidSignal() {
         SignalRequest signalSpec = getSampleSignalSpec();
-        Signal savedSignal = getSampleSignal(signalSpec);
+        Signal savedSignal = ObjectMapperUtil.covertToSignal(signalSpec);
         when(signalService.getSignalById(1)).thenReturn(savedSignal);
         signalHandler.handleSignal(1);
+        Mockito.verify(signalService, Mockito.times(1)).getSignalById(1);
     }
 
     @Test
@@ -80,18 +78,5 @@ public class SingnalHandlerTest {
         actionList.add(reverseAction);
 
         return new SignalRequest(actionList);
-    }
-
-    private Signal getSampleSignal(SignalRequest signalSpec) {
-        Signal signal = new Signal();
-        String jsonActions = null;
-        try {
-            jsonActions = objectMapper.writeValueAsString(signalSpec.actions());
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-        signal.setActions(jsonActions);
-
-        return signal;
     }
 }
